requirements.txt
aiohttp==3.9.5
beautifulsoup4==4.12.3
py-cord==2.6.1
googletrans==3.1.0a0
matplotlib==3.10.0
numpy==2.2.1
python-dotenv==1.0.1
Requests==2.32.3
steam==1.4.4
scipy==1.15.2
aiosqlite==0.19.0
aioftp==0.21.4
aiofiles==23.2.1